Delete AddToDatabase class and replace function CreateDatabase
